--- a/setup.py
+++ b/setup.py
@@ -80,10 +80,7 @@ setup(
     name='h',
     version=versioneer.get_version(),
     description='The Internet. Peer-reviewed.',
-    long_description='\n\n'.join([
-        open('README.rst', 'rt').read(),
-        open('CHANGES.txt', 'rt').read(),
-    ]),
+    long_description=open('README.rst', 'rt').read(),
     classifiers=[
         'Development Status :: 3 - Alpha',
         'Environment :: Console',